requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
websockets>=14.0
openai>=1.60.0
sentence-transformers>=3.4.0
numpy>=1.26.0
pydantic>=2.10.0
python-dotenv>=1.0.0
